Reverted to unrefactored version and added minimal logic for --check flag
