--- src/testing/checking-version/validFallback.Cargo.toml
@@ -1,175 +1,0 @@
-[package]
-name = "runtime"
-version = "1.0.0"
-authors = ["Anonymous"]
-description = "A parachain runtime."
-license = "Unlicense"
-edition = 2021
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[build-dependencies]
-substrate-wasm-builder = { version = "14.0.0", optional = true }
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
-hex-literal = { version = "0.4.1", optional = true }
-log = { version = "0.4.20", default-features = false }
-scale-info = { version = "2.10.0", default-features = false, features = ["derive"] }
-smallvec = "1.11.0"
-
-# Local
-pallet-parachain-template = { version = "0.4.0", default-features = false }
-
-# Substrate
-frame-benchmarking = { version = "25.0.0", default-features = false, optional = true }
-frame-executive = { version = "25.0.0", default-features = false }
-frame-support = { version = "25.0.0", default-features = false }
-frame-system = { version = "25.0.0", default-features = false }
-frame-system-benchmarking = { version = "25.0.0", default-features = false, optional = true }
-frame-system-rpc-runtime-api = { version = "23.0.0", default-features = false }
-frame-try-runtime = { version = "0.31.0", default-features = false, optional = true }
-pallet-aura = { version = "24.0.0", default-features = false }
-pallet-authorship = { version = "25.0.0", default-features = false }
-pallet-balances = { version = "25.0.0", default-features = false }
-pallet-session = { version = "25.0.0", default-features = false }
-pallet-sudo = { version = "25.0.0", default-features = false }
-pallet-timestamp = { version = "24.0.0", default-features = false }
-pallet-transaction-payment = { version = "25.0.0", default-features = false }
-pallet-transaction-payment-rpc-runtime-api = { version = "25.0.0", default-features = false }
-sp-api = { version = "23.0.0", default-features = false }
-sp-block-builder = { version = "23.0.0", default-features = false }
-sp-consensus-aura = { version = "0.29.0", default-features = false }
-sp-core = { version = "25.0.0", default-features = false }
-sp-genesis-builder = { version = "0.4.0", default-features = false }
-sp-inherents = { version = "23.0.0", default-features = false }
-sp-offchain = { version = "23.0.0", default-features = false }
-sp-runtime = { version = "28.0.0", default-features = false }
-sp-session = { version = "24.0.0", default-features = false }
-sp-std = { version = "12.0.0", default-features = false }
-sp-transaction-pool = { version = "23.0.0", default-features = false }
-sp-version = { version = "26.0.0", default-features = false }
-
-# Polkadot
-pallet-xcm = { version = "4.0.0", default-features = false }
-polkadot-parachain-primitives = { version = "3.0.0", default-features = false }
-polkadot-runtime-common = { version = "4.0.0", default-features = false }
-xcm = { version = "4.0.0", package = "staging-xcm", default-features = false }
-xcm-builder = { version = "4.0.1", package = "staging-xcm-builder", default-features = false }
-xcm-executor = { version = "4.0.2", package = "staging-xcm-executor", default-features = false }
-
-# Cumulus
-cumulus-pallet-aura-ext = { version = "0.4.0", default-features = false }
-cumulus-pallet-dmp-queue = { version = "0.4.0", default-features = false }
-cumulus-pallet-parachain-system = { version = "0.4.1", default-features = false, features = ["parameterized-consensus-hook"] }
-cumulus-pallet-session-benchmarking = { version = "6.0.0", default-features = false }
-cumulus-pallet-xcm = { version = "0.4.0", default-features = false }
-cumulus-pallet-xcmp-queue = { version = "0.4.0", default-features = false }
-cumulus-primitives-core = { version = "0.4.0", default-features = false }
-cumulus-primitives-utility = { version = "0.4.1", default-features = false }
-pallet-collator-selection = { version = "6.0.0", default-features = false }
-parachain-info = { version = "0.4.0", package = "staging-parachain-info", default-features = false }
-
-[features]
-default = [ "std" ]
-std = [
-	"codec/std",
-	"cumulus-pallet-aura-ext/std",
-	"cumulus-pallet-dmp-queue/std",
-	"cumulus-pallet-parachain-system/std",
-	"cumulus-pallet-session-benchmarking/std",
-	"cumulus-pallet-xcm/std",
-	"cumulus-pallet-xcmp-queue/std",
-	"cumulus-primitives-core/std",
-	"cumulus-primitives-utility/std",
-	"frame-benchmarking?/std",
-	"frame-executive/std",
-	"frame-support/std",
-	"frame-system-benchmarking?/std",
-	"frame-system-rpc-runtime-api/std",
-	"frame-system/std",
-	"frame-try-runtime?/std",
-	"log/std",
-	"pallet-aura/std",
-	"pallet-authorship/std",
-	"pallet-balances/std",
-	"pallet-collator-selection/std",
-	"pallet-parachain-template/std",
-	"pallet-session/std",
-	"pallet-sudo/std",
-	"pallet-timestamp/std",
-	"pallet-transaction-payment-rpc-runtime-api/std",
-	"pallet-transaction-payment/std",
-	"pallet-xcm/std",
-	"parachain-info/std",
-	"polkadot-parachain-primitives/std",
-	"polkadot-runtime-common/std",
-	"scale-info/std",
-	"sp-api/std",
-	"sp-block-builder/std",
-	"sp-consensus-aura/std",
-	"sp-core/std",
-	"sp-genesis-builder/std",
-	"sp-inherents/std",
-	"sp-offchain/std",
-	"sp-runtime/std",
-	"sp-session/std",
-	"sp-std/std",
-	"sp-transaction-pool/std",
-	"sp-version/std",
-	"substrate-wasm-builder",
-	"xcm-builder/std",
-	"xcm-executor/std",
-	"xcm/std",
-]
-
-runtime-benchmarks = [
-	"cumulus-pallet-parachain-system/runtime-benchmarks",
-	"cumulus-pallet-session-benchmarking/runtime-benchmarks",
-	"cumulus-pallet-xcmp-queue/runtime-benchmarks",
-	"cumulus-primitives-utility/runtime-benchmarks",
-	"frame-benchmarking/runtime-benchmarks",
-	"frame-support/runtime-benchmarks",
-	"frame-system-benchmarking/runtime-benchmarks",
-	"frame-system/runtime-benchmarks",
-	"hex-literal",
-	"pallet-balances/runtime-benchmarks",
-	"pallet-collator-selection/runtime-benchmarks",
-	"pallet-parachain-template/runtime-benchmarks",
-	"pallet-sudo/runtime-benchmarks",
-	"pallet-timestamp/runtime-benchmarks",
-	"pallet-xcm/runtime-benchmarks",
-	"polkadot-parachain-primitives/runtime-benchmarks",
-	"polkadot-runtime-common/runtime-benchmarks",
-	"sp-runtime/runtime-benchmarks",
-	"xcm-builder/runtime-benchmarks",
-	"xcm-executor/runtime-benchmarks",
-]
-
-try-runtime = [
-	"cumulus-pallet-aura-ext/try-runtime",
-	"cumulus-pallet-dmp-queue/try-runtime",
-	"cumulus-pallet-parachain-system/try-runtime",
-	"cumulus-pallet-xcm/try-runtime",
-	"cumulus-pallet-xcmp-queue/try-runtime",
-	"frame-executive/try-runtime",
-	"frame-support/try-runtime",
-	"frame-system/try-runtime",
-	"frame-try-runtime/try-runtime",
-	"pallet-aura/try-runtime",
-	"pallet-authorship/try-runtime",
-	"pallet-balances/try-runtime",
-	"pallet-collator-selection/try-runtime",
-	"pallet-parachain-template/try-runtime",
-	"pallet-session/try-runtime",
-	"pallet-sudo/try-runtime",
-	"pallet-timestamp/try-runtime",
-	"pallet-transaction-payment/try-runtime",
-	"pallet-xcm/try-runtime",
-	"parachain-info/try-runtime",
-	"polkadot-runtime-common/try-runtime",
-	"sp-runtime/try-runtime",
-]
-
-experimental = [ "pallet-aura/experimental" ]
