+++ src/tests.rs
@@ -28,10 +28,11 @@ mod tests {
             .unwrap();
 
         // Call the refactored logic function with the test data
-        let result = crate::update_deps::update_dependencies_impl(
+        let result = crate::update_dependencies_impl(
             &input_cargo_toml_path,
             &crates_versions,
             false,
+            true
         )
         .unwrap();
 
@@ -39,16 +40,23 @@ mod tests {
         assert_eq!(result, Some(expected_cargo_toml.into()));
     }
 
-    async fn check_version_mapping(
+    async fn verify_version_checking(
         version: &str,
         input_cargo_toml_path: &Path,
-    ) -> Result<(), Box<dyn Error>> {
+    ) -> Result<Option<String>, Box<dyn Error>>{
         let crates_versions = get_version_mapping_with_fallback(crate::DEFAULT_GIT_SERVER, version)
             .await
             .unwrap();
 
         // Call the refactored logic function with the test data
-        crate::check_deps::check_dependencies_impl(&input_cargo_toml_path, &crates_versions, false)
+        let result = crate::update_dependencies_impl(
+            &input_cargo_toml_path,
+            &crates_versions,
+            false,
+            false
+        );
+        
+        result
     }
 
     #[tokio::test]
@@ -63,17 +71,47 @@ mod tests {
     }
 
     #[tokio::test]
-    // cargo psvm -v 1.3.0
+    // cargo psvm -v 1.14.0 -c
     // This version doesn't have the Plan.toml file, so it will fallback to Cargo.lock
     // and check if the versions in the local toml file comply with the Cargo.lock file
-    async fn test_check_version_with_fallback_with_valid_format() {
+    async fn test_check_version_without_fallback_passes() {
         let input_cargo_toml_path =
-            Path::new("src/testing/checking-version/validFallback.Cargo.toml");
+            Path::new("src/testing/plan-toml/check.Cargo.toml");
+        let version = "1.14.0";
+
+        let res = verify_version_checking(version, input_cargo_toml_path).await;
+        assert!(res.is_ok());
+        assert!(res.unwrap().is_none());
+    }
+
+    #[tokio::test]
+    // cargo psvm -v 1.4.0 -c
+    // This version doesn't have the Plan.toml file, so it will fallback to Cargo.lock
+    // and check if the versions in the local toml file comply with the Cargo.lock file.
+    // This will fail because the Cargo.toml file has version 1.14.0 while we are checking
+    // against 1.4.0
+    async fn test_check_version_without_fallback_fails_with_incorrect_version() {
+        let input_cargo_toml_path =
+            Path::new("src/testing/plan-toml/check.Cargo.toml");
+        let version = "1.4.0";
+
+        let res = verify_version_checking(version, input_cargo_toml_path).await;
+        assert!(res.is_err());
+        assert_eq!(res.unwrap_err().to_string(), "Dependencies are not up to date");
+    }
+
+    #[tokio::test]
+    // cargo psvm -v 1.3.0 -c
+    // This version doesn't have the Plan.toml file, so it will fallback to Cargo.lock
+    // and check if the versions in the local toml file comply with the Cargo.lock file
+    async fn test_check_version_with_fallback_passes() {
+        let input_cargo_toml_path =
+            Path::new("src/testing/cargo-lock/output.Cargo.toml");
         let version = "1.3.0";
 
-        let res = check_version_mapping(version, input_cargo_toml_path).await;
+        let res = verify_version_checking(version, input_cargo_toml_path).await;
         assert!(res.is_ok());
-        assert_eq!(res.unwrap(), ());
+        assert!(res.unwrap().is_none());
     }
 
     #[tokio::test]
@@ -87,54 +125,54 @@ mod tests {
         verify_version_mapping(version, input_toml_path, expected_output_toml).await;
     }
 
-    #[tokio::test]
-    // cargo psvm -v 1.5.0
-    // This version has the Plan.toml file, so it will not fallback to Cargo.lock
-    // and the corresponding versions of the dependencies in the supplied toml file
-    // will be checked
-    async fn test_check_version_without_fallback_with_valid_format() {
-        let input_cargo_toml_path = Path::new("src/testing/checking-version/valid.Cargo.toml");
-        let version = "1.5.0";
+    // #[tokio::test]
+    // // cargo psvm -v 1.5.0
+    // // This version has the Plan.toml file, so it will not fallback to Cargo.lock
+    // // and the corresponding versions of the dependencies in the supplied toml file
+    // // will be checked
+    // async fn test_check_version_without_fallback_with_valid_format() {
+    //     let input_cargo_toml_path = Path::new("src/testing/checking-version/valid.Cargo.toml");
+    //     let version = "1.5.0";
 
-        let res = check_version_mapping(version, input_cargo_toml_path).await;
-        assert!(res.is_ok());
-        assert_eq!(res.unwrap(), ());
-    }
+    //     let res = check_version_mapping(version, input_cargo_toml_path).await;
+    //     assert!(res.is_ok());
+    //     assert_eq!(res.unwrap(), ());
+    // }
 
-    #[tokio::test]
-    // cargo psvm -v 1.2.0
-    // This version has the Plan.toml file, so it will not fallback to Cargo.lock
-    // and the corresponding versions of the dependencies in the supplied toml file
-    // will be checked. This will result in an error because the toml version is 1.5.0
-    // whereas we are checking for 1.2.0
-    async fn test_check_version_without_fallback_with_valid_format_and_invalid_version() {
-        let input_cargo_toml_path = Path::new("src/testing/checking-version/valid.Cargo.toml");
-        let version = "1.2.0";
+    // #[tokio::test]
+    // // cargo psvm -v 1.2.0
+    // // This version has the Plan.toml file, so it will not fallback to Cargo.lock
+    // // and the corresponding versions of the dependencies in the supplied toml file
+    // // will be checked. This will result in an error because the toml version is 1.5.0
+    // // whereas we are checking for 1.2.0
+    // async fn test_check_version_without_fallback_with_valid_format_and_invalid_version() {
+    //     let input_cargo_toml_path = Path::new("src/testing/checking-version/valid.Cargo.toml");
+    //     let version = "1.2.0";
 
-        let res = check_version_mapping(version, input_cargo_toml_path).await;
-        assert!(res.is_err());
-        assert_eq!(
-            res.unwrap_err().to_string(),
-            "One or More Dependency version mismatch found".to_string()
-        );
-    }
+    //     let res = check_version_mapping(version, input_cargo_toml_path).await;
+    //     assert!(res.is_err());
+    //     assert_eq!(
+    //         res.unwrap_err().to_string(),
+    //         "One or More Dependency version mismatch found".to_string()
+    //     );
+    // }
 
-    #[tokio::test]
-    // cargo psvm -v 1.3.0
-    // This version doesn't have the Plan.toml file, so it will fallback to Cargo.lock
-    // and check if the versions in the local toml file comply with the Cargo.lock file
-    // This will fail because the format of the toml is invalid for checking.
-    async fn test_check_version_with_fallback_with_invalid_format() {
-        let input_cargo_toml_path = Path::new("src/testing/checking-version/invalid.Cargo.toml");
-        let version = "1.3.0";
+    // #[tokio::test]
+    // // cargo psvm -v 1.3.0
+    // // This version doesn't have the Plan.toml file, so it will fallback to Cargo.lock
+    // // and check if the versions in the local toml file comply with the Cargo.lock file
+    // // This will fail because the format of the toml is invalid for checking.
+    // async fn test_check_version_with_fallback_with_invalid_format() {
+    //     let input_cargo_toml_path = Path::new("src/testing/checking-version/invalid.Cargo.toml");
+    //     let version = "1.3.0";
 
-        let res = check_version_mapping(version, input_cargo_toml_path).await;
-        assert!(res.is_err());
-        assert_eq!(
-            res.unwrap_err().to_string(),
-            "Invalid Dependency Format".to_string()
-        );
-    }
+    //     let res = check_version_mapping(version, input_cargo_toml_path).await;
+    //     assert!(res.is_err());
+    //     assert_eq!(
+    //         res.unwrap_err().to_string(),
+    //         "Invalid Dependency Format".to_string()
+    //     );
+    // }
 
     #[tokio::test]
     async fn test_parse_version_mapping_from_plan_toml() {
@@ -245,10 +283,11 @@ source = "registry+https://github.com/rust-lang/crates.io-index"
             );
 
             let input_cargo_toml_path = Path::new("src/testing/plan-toml/input.Cargo.toml");
-            let result = crate::update_deps::update_dependencies_impl(
+            let result = crate::update_dependencies_impl(
                 &input_cargo_toml_path,
                 &crates_versions,
                 false,
+                true
             )
             .unwrap();
